REFACTOR: Move `Ports` out of core submodule for wiring diagrams.
